Support for Papyrus scripts in multiple directories
- Use the path to the file instead of the base name.
- Add the -r flag to the anonymizer call for recursive searching.

Note that the user must pass the relative paths of the scripts to the
current source directory, due to a the Papyrus Compiler limitation.

For example, before add_papyrus is called, the user should enforce
that relative paths is always passed, something along the lines of:

list(TRANSFORM Papyrus_INPUT REPLACE "${CMAKE_CURRENT_SOURCE_DIR}\/*" "")

This is only required if the scripts are structured in folders. Changes
are backwards compatible with existing projects that don't store script
files in sub-folders.